+++ docs/scanning/configuration.md
@@ -278,6 +278,7 @@ Many modules accept their own configuration options. These options have the abil
 | modules.filedownload.base_64_encoded_file      | str      | Stream the bytes of a file and encode them in base 64 for event data.                                                                                                                                                                                                                                                          | false                                                                                                                                                                                                                                                                                                                                                                             |
 | modules.filedownload.extensions                | list     | File extensions to download                                                                                                                                                                                                                                                                                                    | ['bak', 'bash', 'bashrc', 'conf', 'cfg', 'crt', 'csv', 'db', 'sqlite', 'doc', 'docx', 'exe', 'ica', 'indd', 'ini', 'jar', 'key', 'pub', 'log', 'markdown', 'md', 'msi', 'odg', 'odp', 'ods', 'odt', 'pdf', 'pem', 'pps', 'ppsx', 'ppt', 'pptx', 'ps1', 'raw', 'rdp', 'sh', 'sql', 'swp', 'sxw', 'tar', 'tar.gz', 'zip', 'txt', 'vbs', 'wpd', 'xls', 'xlsx', 'xml', 'yml', 'yaml'] |
 | modules.filedownload.max_filesize              | str      | Cancel download if filesize is greater than this size                                                                                                                                                                                                                                                                          | 10MB                                                                                                                                                                                                                                                                                                                                                                              |
+| modules.fingerprintx.skip_common_web           | bool     | Skip common web ports such as 80, 443, 8080, 8443, etc.                                                                                                                                                                                                                                                                        | True                                                                                                                                                                                                                                                                                                                                                                              |
 | modules.fingerprintx.version                   | str      | fingerprintx version                                                                                                                                                                                                                                                                                                           | 1.1.4                                                                                                                                                                                                                                                                                                                                                                             |
 | modules.gitlab.api_key                         | str      | Gitlab access token                                                                                                                                                                                                                                                                                                            |                                                                                                                                                                                                                                                                                                                                                                                   |
 | modules.gowitness.idle_timeout                 | int      | Skip the current gowitness batch if it stalls for longer than this many seconds                                                                                                                                                                                                                                                | 1800                                                                                                                                                                                                                                                                                                                                                                              |
